# api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "0.0.1"
  title: Hello World App
# during dev, should point to your local machine
host: localhost:3000
...
basePath: /
...
securityDefinitions:
  AdminAuth:
    type: basic
  UserAuth:
    type: basic  
  SupplierAuth:
    type: basic
schemes:
  # tip: remove http to make production-grade
  - http
  - https
# format of bodies a client can send (Content-Type)
consumes:
  - application/json
# format of the responses to the client (Accepts)
produces:
  - application/json
paths:
  /user:
    x-swagger-router-controller: user
    get:
      description: get the users' list
      parameters:
        - in: query
          name: top10
          required: false
          type: boolean
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/GetUsersListResponse"
        "401":
          description: Unauthorized
          schema:
            $ref: "#/definitions/UnauthorizedResponse"
        "400":
          description: bad request
          schema:
            $ref: "#/definitions/badRequestResponse"
        "500":
          description: internal server error
          schema:
            $ref: "#/definitions/internalServerErrorRequestResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  
    post:
      description: add a new user to the list
      parameters:
        - name: user
          description:  user information
          in: body
          required: true
          schema:
            $ref: "#/definitions/user"
      responses:
        "201":
          description: created
          schema:
            $ref: "#/definitions/GetuserResponse"
        "400":
          description: bad request
          schema:
            $ref: "#/definitions/badRequestResponse"
        "500":
          description: internal server error
          schema:
            $ref: "#/definitions/internalServerErrorRequestResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
      security: []
...
  /order:
    x-swagger-router-controller: order
    get:
      description: get the order list for the logged in user or return the whole orders list if the requester is an admin
      parameters:
        - in: query
          name: status
          type: integer
          description: /order?status=pending
      # define the type of response for Success "200" and Error
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/GetOrdersListResponse"
        "401":
          description: Unauthorized
          schema:
            $ref: "#/definitions/UnauthorizedResponse"
        "400":
          description: bad request
          schema:
            $ref: "#/definitions/badRequestResponse"
        "500":
          description: internal server error
          schema:
            $ref: "#/definitions/internalServerErrorRequestResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
      security:
         - AdminAuth: []
         - UserAuth: []
    post:
      description: add a new order to the cart
      # movie info to be stored
      parameters:
        - name: order
          description:  new order
          in: body
          required: true
          schema:
            $ref: "#/definitions/order"
      responses:
        "201":
          description: created
          schema:
            $ref: "#/definitions/GetOrderResponse"
        "401":
          description: Unauthorized
          schema:
            $ref: "#/definitions/UnauthorizedResponse"
        "500":
          description: internal server error
          schema:
            $ref: "#/definitions/internalServerErrorRequestResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
      security:
         - AdminAuth: []
         - UserAuth: []
...
  /order/{id}/confirmation:
    get:
      description: get confirmation status for the order to check whether is was confirmed or not
      parameters:
        - name: id
          type: string
          in: path
          required: true
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/Confirmation"
        "401":
          description: Unauthorized
          schema:
            $ref: "#/definitions/UnauthorizedResponse"
        "500":
          description: internal server error
          schema:
            $ref: "#/definitions/internalServerErrorRequestResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
      security:
         - AdminAuth: []
         - UserAuth: []
    patch:
      description: confirm given order
      parameters:
        - name: id
          type: string
          in: path
          required: true
        - name: confirmation
          description:  post a confirmation for the order
          in: body
          required: true
          schema:
            $ref: "#/definitions/Confirmation"
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/GetOrderResponse"
        "400":
          description: bad request
          schema:
            $ref: "#/definitions/badRequestResponse"
        "401":
          description: Unauthorized
          schema:
            $ref: "#/definitions/UnauthorizedResponse"
        "500":
          description: internal server error
          schema:
            $ref: "#/definitions/internalServerErrorRequestResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
      security:
         - AdminAuth: []
         - UserAuth: []
...
  /process-table:
    x-swagger-router-controller: process-table
    get:
      description: get process table
      # define the type of response for Success "200" and Error
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/GetTableResponse"
        "401":
          description: Unauthorized
          schema:
            $ref: "#/definitions/UnauthorizedResponse"
        "500":
          description: internal server error
          schema:
            $ref: "#/definitions/internalServerErrorRequestResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
      security:
         - AdminAuth: []
         - UserAuth: []
    post:
      description: add a new stage to the process table
      parameters:
        - name: cat
          description:  cat info
          in: body
          required: true
          schema:
            $ref: "#/definitions/stage"
      responses:
        "201":
          description: created
          schema:
            $ref: "#/definitions/GetStageResponse"
        "400":
          description: bad request
          schema:
            $ref: "#/definitions/badRequestResponse"
        "401":
          description: Unauthorized
          schema:
            $ref: "#/definitions/UnauthorizedResponse"
        "500":
          description: internal server error
          schema:
            $ref: "#/definitions/internalServerErrorRequestResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
      security:
         - AdminAuth: []
  /Process-table/{id}:
    get:
      description: get information about stage of the process table details
      # define the type of response for Success "200" and Error
      parameters:
        - name: id
          type: string
          in: path
          required: true
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/GetStageResponse"
        "401":
          description: Unauthorized
          schema:
            $ref: "#/definitions/UnauthorizedResponse"
        "500":
          description: internal server error
          schema:
            $ref: "#/definitions/internalServerErrorRequestResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
      security:
         - AdminAuth: []
         - UserAuth: []
    put:
      description: update one stage information
      # movie info to be stored
      parameters:
        - name: id
          type: string
          in: path
          required: true
        - name: stage
          description: stage information
          in: body
          required: true
          schema:
            $ref: "#/definitions/stage"
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/GetStageResponse"
        "400":
          description: bad request
          schema:
            $ref: "#/definitions/badRequestResponse"
        "401":
          description: Unauthorized
          schema:
            $ref: "#/definitions/UnauthorizedResponse"
        "500":
          description: internal server error
          schema:
            $ref: "#/definitions/internalServerErrorRequestResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
      security:
         - AdminAuth: []
    delete:
      description: delete a stage from the process table
      # define the parameters
      parameters:
  
        - name: id
          description: stage id
          type: string
          in: path
          required: true
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/GeneralDeletedddSuccessfullyResponse"
        "401":
          description: Unauthorized
          schema:
            $ref: "#/definitions/UnauthorizedResponse"
        "404":
          description: Not found
          schema:
            $ref: "#/definitions/notFoundResponse"
        "500":
          description: internal server error
          schema:
            $ref: "#/definitions/internalServerErrorRequestResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
      security:
         - AdminAuth: []
...
definitions:
  ErrorResponse:
    required:
      - message
    properties:
      message:
        type: string
  UnauthorizedResponse:
    required:
      - message
    properties:
      message:
        type: string
  badRequestResponse:
    required:
      - message
    properties:
      message:
        type: string
  internalServerErrorRequestResponse:
    required:
      - message
    properties:
      message:
        type: string
  notFoundResponse:
    required:
      - message
    properties:
      message:
        type: string
...
  GeneralDeletedddSuccessfullyResponse:
    type: object
    properties:
      success:
        type: number
        description: returns 1 if successful
      description:
        type: string
  GetUsersListResponse:
    properties:
      user: 
        type: object
        properties:
             _links: 
                type: object
                properties:
                    self:
                      type: object
                      properties:
                        href:
                          type: string
             _embedded:
              type: object
              properties:
                 usersList:
                    type: array
                    items:
                      type: object
                      properties:
                          _links: 
                              type: object
                              properties:
                                  self:
                                    type: object
                                    properties:
                                      href:
                                        type: string
                          _id:
                            type: string
                          username:
                            type: string
  GetuserResponse:
    type: object
    properties:
      _links:
        type: object
        properties:
          self:
            type: object
            properties:
              href:
                type: string
      _embedded:
        type: object
        properties:
          orders:
            type: array
            items:
                type: object
                properties:
                  _links:
                    type: object
                    properties:
                      self:
                        type: object
                        properties:
                          href:
                            type: string
                  _id: 
                    type: string
          totalSales:
            type: number
      _id:
        type: string
        description: user unique id
      username:
        type: string
        description: user name
      password:
        type: string
        description: user password
      email:
        type: string
        description: user email
      createdAt:
       type: string
      updatedAt:
       type: string
      __v:
       type: number
  user:
    type: object
    properties:
      username:
        type: string
        description: user name
      password:
        type: string
        description: user password
      email:
        type: string
        description: user email
    required:
      - username
      - password
      - email  
...
  order:
          required:
           - stamping
           - dyeing
           - quantity
           - deliveryDate
           - delivaryPlace
  
          type: object
          properties:
            stamping:
              type: string
            stampingImg:
              type: string
            dyeing:
              type: string
            dyeingColor:
              type: string
            quantity:
              type: number
            deliveryDate:
              type: string
            delivaryPlace:
              type: object
              properties:
                street:
                  type: string
                city:
                  type: string
                state:
                  type: string
                country:
                  type: string
                zipCode: 
                  type: number
 
            fabricSpec:
              type: array
              items: 
                type: object
                properties:
                    code:
                      type: string
                    NE: 
                      type: number
                    cables: 
                      type: number
                    filament: 
                      type: number
                    TPM: 
                      type: number
                    percintage: 
                      type: number
  GetOrderResponse:
    properties:
      _links:
         type: object
         properties:
              self:
                type: object
                properties:
                    href: 
                      type: string
              owner:
                type: object
                properties:
                    href: 
                      type: string
                    userName:
                      type: string
      _embedded:
         type: object
         properties:
           invoice:
             type: object
             properties:
               self:
                 type: object
                 properties:
                   href: 
                     type: string
           costDetails:
             type: object
             properties:
               self:
                 type: object
                 properties:
                   href:
                     type: string
           confirmation:
             type: object
             properties:
               self:
                 type: object
                 properties:
                   href:
                     type: string
           stampingImg:
             type: object
             properties:
               self:
                 type: object
                 properties:
                   href:
                     type: string

      id:
        type: string
      orderDate:
        type: string
      stamping:
        type: boolean
      dyeing:
        type: boolean
      dyeingColor:
        type: string
      quantity:
        type: number
      deliveryDate:
        type: string
      delivaryPlace:
        type: object
        properties:
          street:
            type: string
          city:
            type: string
          state:
            type: string
          country:
            type: string
          zipCode: 
            type: number
      status:
        type: string
      fabricSpec:
        type: array
        items: 
          type: object
          properties:
              code:
                type: string
              NE: 
                type: number
              cables: 
                type: number
              filament: 
                type: number
              TPM: 
                type: number
              percintage: 
                type: number
      confirmed:
        type: boolean
      totalCost:
        type: number
      currency: 
        type: string
  Confirmation:
    properties:
      confirmed:
        type: boolean
  GetOrdersListResponse:
    properties:
      _links:
           type: object
           properties:
              self:
                type: object
                properties:
                    href: 
                      type: string

      _embedded:
        type: object
        properties:
          orders:
            type: array
            items: 
              type: object
              properties:
                _links:
                  type: object
                  properties:
                    self:
                      type: object
                      properties:
                        href:
                          type: string
                    owner:
                      type: object
                      properties:
                          href: 
                            type: string
                          userName:
                            type: string
                id:
                  type: string
                date:
                  type: string
                status:
                  type: string
          totalSales:
            type: number
...
  stage:
    required:
        - name
        - description
        - order
        - price
        - currency
        - optional
    type: object
    properties:
        name:
          type: string
        description:
          type: string
        order:
          type: number
        price:
          type: number
        currency:
          type: number
        optional:
          type: string
  GetStageResponse:
     type: object
     properties:
      _links:
         type: object
         properties:
           self:
             type: object
             properties:
               href:
                type: string
           process-table:
             type: object
             properties:
               href:
                type: string
      name:
        type: string
      description:
        type: string
      id:
        type: string
      order:
        type: number
      price:
        type: number
      currency:
        type: number
      optional:
        type: string
  GetTableResponse:
     type: object
     properties:
       _links:
           type: object
           properties:
             self:
               type: object
               properties:
                 href:
                  type: string
       _embedded:
          type: object
          properties:
            stages:
              type: array
              items:
                 type: object
                 properties:
                    _links:
                       type: object
                       properties:
                         self:
                           type: object
                           properties:
                             href:
                               type: string
                    stageName: 
                       type: string
...